console output is routed again to REPL
